requirements.txt
# Existing dependencies
openai>=1.0.0
httpx>=0.24.0
python-dotenv>=1.0.0

# New dependencies for multi-agent system
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
python-multipart>=0.0.6

